+++ src/main/java/de/samply/security/WebAppSecurityConfig.java
@@ -1,0 +1,84 @@
+package de.samply.security;
+
+import org.camunda.bpm.webapp.impl.security.auth.ContainerBasedAuthenticationFilter;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingClass;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.web.context.request.RequestContextListener;
+import org.springframework.web.filter.ForwardedHeaderFilter;
+
+import java.util.Collections;
+
+import static org.springframework.security.config.Customizer.withDefaults;
+import static org.springframework.security.web.util.matcher.AntPathRequestMatcher.antMatcher;
+
+/**
+ * Camunda Web application SSO configuration for usage with KeycloakIdentityProviderPlugin.
+ */
+@ConditionalOnMissingClass("org.springframework.test.context.junit.jupiter.SpringExtension")
+@EnableWebSecurity
+@Configuration
+public class WebAppSecurityConfig {
+
+    @Value("${camunda.bpm.webapp.application-path}")
+    private String camundaApplicationPath;
+
+    private String fetchPath(String suffix){
+        return (camundaApplicationPath != null) ? camundaApplicationPath + suffix : suffix;
+    }
+
+    @Bean
+    @Order(1)
+    public SecurityFilterChain httpSecurity(HttpSecurity http) throws Exception {
+        return http
+                .csrf(csrf -> csrf
+                        .ignoringRequestMatchers(antMatcher(fetchPath("/api/**")), antMatcher(fetchPath("/engine-rest/**"))))
+                .authorizeHttpRequests(authorize -> authorize
+                        .requestMatchers(
+                                antMatcher(fetchPath("/assets/**")),
+                                antMatcher(fetchPath("/app/**")),
+                                antMatcher(fetchPath("/api/**")),
+                                antMatcher(fetchPath("/lib/**")))
+                        .authenticated()
+                        .anyRequest()
+                        .permitAll())
+                .oauth2Login(withDefaults())
+                .build();
+    }
+
+    @SuppressWarnings({"rawtypes", "unchecked"})
+    @Bean
+    public FilterRegistrationBean containerBasedAuthenticationFilter() {
+
+        FilterRegistrationBean filterRegistration = new FilterRegistrationBean();
+        filterRegistration.setFilter(new ContainerBasedAuthenticationFilter());
+        filterRegistration.setInitParameters(Collections.singletonMap("authentication-provider", "de.samply.security.KeycloakAuthenticationProvider"));
+        filterRegistration.setOrder(101); // make sure the filter is registered after the Spring Security Filter Chain
+        filterRegistration.addUrlPatterns(fetchPath("/app/*"));
+        return filterRegistration;
+    }
+
+    // The ForwardedHeaderFilter is required to correctly assemble the redirect URL for OAUth2 login.
+    // Without the filter, Spring generates an HTTP URL even though the container route is accessed through HTTPS.
+    @Bean
+    public FilterRegistrationBean<ForwardedHeaderFilter> forwardedHeaderFilter() {
+        FilterRegistrationBean<ForwardedHeaderFilter> filterRegistrationBean = new FilterRegistrationBean<>();
+        filterRegistrationBean.setFilter(new ForwardedHeaderFilter());
+        filterRegistrationBean.setOrder(Ordered.HIGHEST_PRECEDENCE);
+        return filterRegistrationBean;
+    }
+
+    @Bean
+    @Order(0)
+    public RequestContextListener requestContextListener() {
+        return new RequestContextListener();
+    }
+
+}