+++ src/main/java/de/samply/pseudonymisation/Patient.java
@@ -3,8 +3,10 @@ package de.samply.pseudonymisation;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.xml.bind.annotation.XmlRootElement;
+
 import java.util.List;
-import javax.xml.bind.annotation.XmlRootElement;
+
 
 @XmlRootElement
 @JsonInclude(Include.NON_NULL)